v1.7.5 Cleaned up code to avoid compile warnings. Sopwith now builds cleanly using Clang 3.0 and GCC 4.6.
Added custom keys option. Controls can be altered by
adjusting the values in the ~/.sopwith/keys file. On
Windows this file should be located in the user's profile
under "AppData/Local/.sopwith/keys".

[This has been modified as part of the 2022 Git conversion to squash the Id
 and Log CVS tags]

+++ src/swend.c
@@ -37,7 +37,8 @@ static void swreport()
 {
 	puts("\r\nEnd of game statictics\r\n\r\n");
 	puts("Objects used: ");
-	dispd(((int) objsmax - (int) objtop + 1) / sizeof(OBJECTS), 0);
+	// dispd(((int) objsmax - (int) objtop + 1) / sizeof(OBJECTS), 0);
+        dispd( (int) (objsmax - objtop + 1) / sizeof(OBJECTS), 0);
 	puts("\r\n");
 }
